+++ src/user-action/createTargetedUserAction.test.ts
@@ -6,7 +6,7 @@ import { GravityDocument, UserActionType } from '../types'
 import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 import { createClickUserAction, createKeyDownUserAction, createKeyUpUserAction } from '../test-utils/userActions'
 
-describe('action', () => {
+describe('user action', () => {
   let document: GravityDocument
 
   beforeEach(() => {
@@ -17,23 +17,23 @@ describe('action', () => {
 
   describe('createTargetedUserAction', () => {
     it('returns the specified "type"', () => {
-      const { element } = createElementInJSDOM('<button>Click Me</button>', 'button')
+      const { element, domWindow } = createElementInJSDOM('<button>Click Me</button>', 'button')
 
-      const action = createClickUserAction(element)
+      const action = createClickUserAction(element, 0, 0, domWindow.document)
 
       expect(action.type).toEqual(UserActionType.Click)
     })
 
     it('returns location data', () => {
-      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+      const { element, domWindow } = createElementInJSDOM('<div>Click Me</div>', 'div')
 
-      const action = createClickUserAction(element)
+      const action = createClickUserAction(element, 0, 0, domWindow.document)
 
       expect(action?.location).toEqual(location())
     })
 
     it('returns document data', () => {
-      const { element } = createElementInJSDOM(
+      const { element, domWindow } = createElementInJSDOM(
         '<html lang="en">' +
           ' <head>' +
           '   <title></title>' +
@@ -45,15 +45,15 @@ describe('action', () => {
         'html',
       )
 
-      const action = createClickUserAction(element)
+      const action = createClickUserAction(element, 0, 0, domWindow.document)
 
       expect(action?.document.title).toEqual(document.title)
     })
 
     it('returns viewport data', () => {
-      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+      const { element, domWindow } = createElementInJSDOM('<div>Click Me</div>', 'div')
 
-      const action = createClickUserAction(element)
+      const action = createClickUserAction(element, 0, 0, domWindow.document)
 
       expect(action?.viewportData).toEqual(viewport())
     })
@@ -63,67 +63,79 @@ describe('action', () => {
       vi.useFakeTimers()
       vi.setSystemTime(Date.parse('2022-05-12'))
 
-      const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+      const { element, domWindow } = createElementInJSDOM('<div>Click Me</div>', 'div')
 
-      const action = createClickUserAction(element)
+      const action = createClickUserAction(element, 0, 0, domWindow.document)
 
       expect(action?.recordedAt).toEqual(now)
     })
 
     describe('target data embeds...', () => {
       it('tag name', () => {
-        const { element } = createElementInJSDOM('<div>Click Me</div>', 'div')
+        const { element, domWindow } = createElementInJSDOM('<div>Click Me</div>', 'div')
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action.target.element).toEqual('div')
       })
 
       it('type is recorded if element has the type attributes', () => {
-        const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+        const { element, domWindow } = createElementInJSDOM(
+          '<input type="text" data-testid="userName" class="size-lg"/>',
+          'input',
+        )
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action?.target.type).toEqual('text')
       })
 
       it('selector', () => {
-        const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+        const { element, domWindow } = createElementInJSDOM(
+          '<input type="text" data-testid="userName" class="size-lg"/>',
+          'input',
+        )
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action.target?.selector).toEqual('.size-lg')
       })
 
       it('value not recorded if input is a text box', () => {
-        const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+        const { element, domWindow } = createElementInJSDOM(
+          '<input type="text" data-testid="userName" class="size-lg"/>',
+          'input',
+        )
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action.target?.value).toBeUndefined()
       })
 
       it('"true" is recorded if input is a checked checkbox', () => {
-        const { element } = createElementInJSDOM('<input type="checkbox" class="size-lg" checked/>', 'input')
+        const { element, domWindow } = createElementInJSDOM('<input type="checkbox" class="size-lg" checked/>', 'input')
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action.target?.value).equals('true')
       })
 
       it('"false" is recorded if input is a checked checkbox', () => {
-        const { element } = createElementInJSDOM('<input type="checkbox" class="size-lg"/>', 'input')
+        const { element, domWindow } = createElementInJSDOM('<input type="checkbox" class="size-lg"/>', 'input')
 
-        const action = createClickUserAction(element)
+        const action = createClickUserAction(element, 0, 0, domWindow.document)
 
         expect(action.target?.value).equals('false')
       })
     })
 
     it('pointer coordinates data when the event is a click', () => {
-      const { element } = createElementInJSDOM('<input type="text" data-testid="userName" class="size-lg"/>', 'input')
+      const { element, domWindow } = createElementInJSDOM(
+        '<input type="text" data-testid="userName" class="size-lg"/>',
+        'input',
+      )
 
-      const action = createClickUserAction(element, 12, 34)
+      const action = createClickUserAction(element, 12, 34, domWindow.document)
 
       const eltBounds = element?.getBoundingClientRect()
 
@@ -138,8 +150,8 @@ describe('action', () => {
     })
 
     it('key data when the event is a keyup', () => {
-      const { element } = createElementInJSDOM('<div/>', 'div')
-      const action = createKeyUpUserAction(element, 'Shift', 'ShiftLeft')
+      const { element, domWindow } = createElementInJSDOM('<div/>', 'div')
+      const action = createKeyUpUserAction(element, 'Shift', 'ShiftLeft', domWindow.document)
       expect(action.userActionData).toEqual({
         key: 'Shift',
         code: 'ShiftLeft',
@@ -147,8 +159,8 @@ describe('action', () => {
     })
 
     it('key data when the event is a keydown', () => {
-      const { element } = createElementInJSDOM('<div/>', 'div')
-      const action = createKeyDownUserAction(element, 'Shift', 'ShiftLeft')
+      const { element, domWindow } = createElementInJSDOM('<div/>', 'div')
+      const action = createKeyDownUserAction(element, 'Shift', 'ShiftLeft', domWindow.document)
       expect(action.userActionData).toEqual({
         key: 'Shift',
         code: 'ShiftLeft',